==> notebooks/1. PreProcessing.ipynb <==
import pandas as pd
from sklearn.preprocessing import LabelEncoder
from sklearn.cluster import KMeans, AgglomerativeClustering
from sklearn.mixture import GaussianMixture
from sklearn.metrics import silhouette_score, davies_bouldin_score, calinski_harabasz_score

df = pd.read_excel("/home/abdeldjalil-hani/Desktop/emergency-sorting-system/data-sets/0. DATA.xlsx")

df.info()

def preprocess_data(df):
    df['Risk Factors'] = df['Risk Factors'].fillna('No Risk Factor')
    df['gender'] = df['gender'].fillna(1)
    df = df.drop(columns=['ID'])

    df[['blood_pressure', 'heart_pressure']] = df['Blood Pressure (mmHg)'].str.split('/', expand=True)
    df['blood_pressure'] = df['blood_pressure'].astype(int)
    df['heart_pressure'] = df['heart_pressure'].astype(int)
    df = df.drop(columns=['Blood Pressure (mmHg)'])

    label_encoders = {}
    categorical_columns = ['Residence_type', 'smoking_status', 'Symptom', 'Consciousness', 'Risk Factors', 'Massive Bleeding', 'Respiratory Distress']

    for col in categorical_columns:
        le = LabelEncoder()
        df[col] = le.fit_transform(df[col])
        label_encoders[col] = le
        print(f"Encoding for '{col}':")
        for cls, code in zip(le.classes_, le.transform(le.classes_)):
            print(f"  {cls} --> {code}")
        print()

    return df, label_encoders

df_processed, encoders = preprocess_data(df)
df_processed.to_excel("/home/abdeldjalil-hani/Desktop/emergency-sorting-system/data-sets/1. PreProcessed DATA.xlsx", index=False)

df_processed.info()
